requirements.txt
pymupdf
pillow
transformers
torch
